Hybrid using new features

+++ get_features.py
@@ -31,4 +31,4 @@ def get_features():
     target = feature_df['warfarin']
     num_cols = get_num_cols(feature_df).drop('warfarin')
     #import ipdb; ipdb.set_trace()
-    return feature_df[num_cols].fillna(-1), target
+    return feature_df[num_cols].fillna(-1).astype(float), target
